pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moba-draft-agent"
version = "0.1.0"
description = "Agente de conversação para pick/ban em League of Legends"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
dependencies = [
    "pyyaml>=6.0.1",
    "openai>=1.40.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "python-dotenv>=1.0.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
